+++ CHANGELOG.md
@@ -13,6 +13,11 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+### Fixed
+- Event query by isDelivered
+- Event async update
+### Added
+- workspace.update() to allow Workspace parameter updates
 
 ## [2.5.2] - 2021-04-23
 ### Fixed